+++ scripts/check.editorconfig.sh
@@ -1,6 +1,6 @@
 # Подсчет кол-ва строк кода в исходных файлах
 
-source "$(dirname $0)"/.config.sh
+source "$(dirname $0)"/../bash.config.sh
 
 echo -e "${YELLOW}Проверяю editorconfig...${WHITE}"