+++ scripts/i.sh
@@ -1,6 +1,6 @@
 # Устанавливает зависимости
 
-source "$(dirname $0)"/.config.sh
+source "$(dirname $0)"/../bash.config.sh
 
 echo -e "${YELLOW}Устанавливаю зависимости...${WHITE}"