+++ scripts/lint.eslint.sh
@@ -1,6 +1,6 @@
 # Launch scripts linter
 
-source "$(dirname $0)"/.config.sh
+source "$(dirname $0)"/../bash.config.sh
 
 OPTION=${1}