+++ scripts/lint.htmllint.sh
@@ -1,6 +1,6 @@
 # Launch HTML linter
 
-source "$(dirname $0)"/.config.sh
+source "$(dirname $0)"/../bash.config.sh
 
 FILES='**/*.html'