+++ scripts/lint.markdownlint.sh
@@ -1,6 +1,6 @@
 # Launch linter for .md files
 
-source "$(dirname $0)"/.config.sh
+source "$(dirname $0)"/../bash.config.sh
 
 OPTION=${1}