+++ scripts/lint.npmlint.sh
@@ -1,6 +1,6 @@
 # Launch NPM linter
 
-source "$(dirname $0)"/.config.sh
+source "$(dirname $0)"/../bash.config.sh
 
 PACKAGE_FILES='. ./packages'
 LOCK_FILE='package-lock.json'