+++ scripts/lint.sh
@@ -1,6 +1,6 @@
 # Запускает линтеры
 
-source "$(dirname $0)"/.config.sh
+source "$(dirname $0)"/../bash.config.sh
 
 OPTION=${1}