+++ scripts/lint.stylelint.sh
@@ -1,6 +1,6 @@
 # Launch styles linter
 
-source "$(dirname $0)"/.config.sh
+source "$(dirname $0)"/../bash.config.sh
 
 OPTION=${1}