+++ scripts/release.sh
@@ -1,6 +1,6 @@
 # Релиз пакета (коммит в npm и git)
 
-source "$(dirname $0)"/.config.sh
+source "$(dirname $0)"/../bash.config.sh
 
 VERSION=${1:-patch} # patch | minor | major