+++ scripts/semver.sh
@@ -1,6 +1,6 @@
 # Обновляет версию пакета
 
-source "$(dirname $0)"/.config.sh
+source "$(dirname $0)"/../bash.config.sh
 
 VERSION=${1:-patch} # <major | minor | patch | premajor | preminor | prepatch | prerelease>